+++ src/test/resources/inspections/codeStyle/ParameterDefaultsNullInspection/default.php
@@ -100,3 +100,17 @@ $dummy = new class
     {
     }
 };
+
+class DummyE
+{
+    function dummy(int $a = 123)
+    {
+    }
+}
+
+$dummy = new class
+    extends DummyE {
+    function dummy(int $a = 123)
+    {
+    }
+};